ZTerminal: Rework the signal handling when suspending or resuming.
These signals can interrupt in the middle of e.g. repainting. Avoid the
rest of the mainline program emitting anything by replacing the file handle
by a fd pointing to /dev/null. Restore it back to the real terminal in the
event loop where it's safer to assume nothing is writing to it.
Also use termpaint's unpause to restore the terminal to expected state.