ZValuePtr: Constrain inplace constructor more to avoid incomplete type errors.
For example clang 7 using stdc++ fails when ZValuePtr is used with an
incomplete type while 'while declaring the implicit copy constructor'
for a class that contains ZValuePtr.

As this inplace/forwarding constructor should never be called with
ZValuePtr as argument* just disable this overload in this case.

* recursive structures might, but those should have no use for ZValuePtr
as this is primarily intended as pimpl helper class and recursion
through such a pointer is only possible in the implemenatation which
does not need ZValuePtr.