ZTerminal: Handle broken terminal connections and add signal terminalConnectionLost for applications to handle this condition.
In the read socket notifier detect when the terminal is no longer
connected and stop polling (which would lead to 100% CPU usage).

If this condition occurs allow the application to override further
handling using the terminalConnectionLost signal. If that signal is not
connected fall back to terminating the currently running event loop.

For this signal to be actually sent setup a signal handler for SIGHUP
that just ignores the signal. This stops SIGHUP immediatly terminating
the application and allows the read code to handle the terminal
disconnect/hangup condition. As a side effect spurious SIGHUP are now
ignored but that should not be a problem.