ZWindowFacet:: Add member variable and setter for manually placed and (centered only) auto placement.
ZWindow needs setManuallyPlaced for move/resize operations so might
as well move this here to avoid making ZBasicWindowFacet a defacto
required base class.

As this allows setting manually placed to false, also make the autoPlace
implementation do something. Centering seems to be the best balance
of complexity and usefulness so that what we are doing here.