ZTerminal,ZWidget: Defer attaching widget tree to terminal until auto detection is finished. Fudge focus handling for unattached widgets.
Character measurements are wrong before terminal auto detection is
finished. Thus widgets run the risk of getting wrong initial data with no
clear indication when it is safe to use character measurements.

This reuses the code paths for attaching a widget tree to a terminal to
avoid this special case.

Having the widget tree attach late to the terminal complicates initial
focus handling when focus only works when fully attached to a terminal.
To avoid this add a workaround that allows setting focus before attaching
to a terminal.