Layouts: Implement removeWidgetRecursively and removal of child layouts.
When a widget is reparented away the wrapping layout item is always
removed, leaving the widget itself intact.
For layouts that are reparented away we leave them be as either the new
parent or the code that reparented them is now responsible for their
lifetime.

On destruction the child layouts are already destroyed by the usual parent
child relation so layouts can be left alone there too.

In case a layout item is replaced by a new layout item the previous layout
item is destroyed unless it is a layout. As layouts are lifetime managed
by the parent child relation we assume that they will eventually either
be destroyed by their parent or reparented away and someone else assumes
responsibility for their lifetime.