ZTextLineRef::draw: Organize format ranges to avoid O(text runs * ranges) intersection checking.
When having many format ranges and using one of the modes that produce many
text runs, the previous brute force search of every format range for every
text run was very slow.

As the ordering of the ranges matters (the last one that covers a character
wins) simply sorting the ranges in not possible.
As text runs are already sorted, a binary search can be used to match a
format range to the first text run and linear searching to find all further
intersecting text runs. Format ranges are then stored in a temporary buffer
with that information and preserving their order.
Using a stable sort on the text run index for each of these the format
ranges are rearranged per text run, and the actual rendering code just
needs to iterate once over the result.

This reduces complexity from O(text runs * ranges) to
O(text runs + (ranges * log(text runs)) + (a * ranges * log(a * ranges)))
where a is the average number of text runs a range intersects.
Assuming a is close to 1 or the whole term is small, this is a nice speedup.

Testing with a 3646 char json text and syntax highlighting yields a 13 fold
speedup.