ZDocument: Fix crash in regular expression search for '\n'.
The implicit line break after the last line of the document is not
addressable with cursors, so the code translating the offset in the search
buffer to a cursor position would do an out of bounds read on the line
vector.

The non regular expression search doesn't find this position either, so
just not add the '\n' to the last line in the regular expression search
buffer for now.