ZWidget: Change mode of access to children for z-reordering.
The (re)stacking code in ZWidget uses undocumented Qt internals to
reorder QObject children.
With qt5 it was fairly safe to assume that nothing would be changing
with these qt internals anymore.

Currently the restacking code in ZWidget accesses the list of children
by referencing through d_ptr which relies on the names not changing.
This of course is not guaranteed.
But QObject::children is an inline function that returns a readonly
reference to the name we need and by being an inline function the ABI
stability guarantees restrict how much things can change.

Thus instead of going though d_ptr, instead const_cast the return of
QObject::children for reordering.

The only change that qt could do in the qt6 series, that doesn't break the
ABI, i can think of would be to use a cached copy there instead of the
real working version of the children, which seems not too likely.

But we still should ask for a stable api for our usecase.